Prepare for Mk1 support by moving common surface code into a base class and moving protocol specific code into a separate class and source file.
